Add files via upload
Hi all,

These are the code files I currently have in my teamcode folder. When you want to use these files, don't forget to put them into teamcode folder, not example folder. Unless you do it you won't be able to build an app containing the codes.

You are free to make an edit to the code and re-upload it, but please let me know and upload it as a branch so that I can track the changes.